Fix issues with circular dependencies between models.
