+++ api/namex/models/__init__.py
@@ -12,12 +12,13 @@ from marshmallow.exceptions import ValidationError
 
 from .applicant import Applicant, ApplicantSchema
 from .comment import Comment, CommentSchema, NameCommentSchema
+from .state import State, StateSchema
 from .event import Event
 from .nwpta import PartnerNameSystem, PartnerNameSystemSchema
 from .name import Name, NameSchema
 from .request import Request, RequestsSchema, RequestsHeaderSchema, RequestsSearchSchema
 from .user import User, UserSchema
-from .state import State, StateSchema
+
 from .decision_reason import DecisionReason
 from .nro_sync_tables import NRONamesSyncJob, NRONamesSyncJobDetail, NRONamesSyncJobStatus
 from .admin_tables import DecisionReasonAudit, RestrictedConditionAudit
